refactor: add styles for collection components
