+++ src/app/collection/list/collection/styles.module.scss
@@ -1,0 +1,12 @@
+.root {
+  position: relative;
+  width: 272px;
+  min-height: 400px;
+  max-height: 400px;
+}
+
+.description {
+  min-height: 125px;
+  max-height: 125px;
+  overflow-y: auto;
+}
