+++ src/app/collection/list/collectionTable/styles.module.scss
@@ -1,0 +1,5 @@
+.root {
+  justify-content: center;
+  gap: 1rem;
+  margin-top: 2rem;
+}
